Implement interrupt on Windows
Sends CTRL+C currently, but easily configurable
to send CTRL+BREAK, should we need that.

This is a bit difficult to implement on windows,
because you need to attach the console of the
target process, which means that the calling
process will lose its console. So we create a
subprocess just for this, this is apparently
the only clean solution. We run the child
process synchronously, it is fairly fast, about
~ 0.05s.